rafi_assistant/requirements.txt
python-telegram-bot==22.5
fastapi==0.128.4
uvicorn[standard]==0.39.0
twilio==9.10.1
elevenlabs==2.34.0
openai==2.17.0
anthropic==0.79.0
deepgram-sdk==5.3.2
google-api-python-client==2.189.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0
supabase==2.27.3
httpx==0.28.1
apscheduler==3.11.2
pyyaml==6.0.3
pydantic==2.12.5
cryptography==46.0.4
python-json-logger==4.0.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1
mypy==1.19.1
